requirements.txt
streamlit
youtube-transcript-api
google-genai
git+https://github.com/Vu0401/zswarm.git
